solution/pyproject.toml
[build-system]
requires = ["setuptools==80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "antifraud"
requires-python = ">=3.13.11"
version = "1.0.0"
readme = 'README.md'
description = "FastAPI backend for AntiFraud project"
dependencies = [
    "asyncpg==0.31.0",
    "dishka==1.7.2",
    "fastapi==0.128.0",
    "Jinja2==3.1.6",
    "alembic==1.17.2",
    "uvicorn[standard]==0.40.0",
    "SQLAlchemy==2.0.45",
    "redis==7.1.0",
    "rust-just==1.45.0",
    "adaptix==3.0.0b11",
    "pydantic==2.12.5",
    "pydantic-extra-types==2.10.6",
    "email-validator==2.3.0",
    "argon2-cffi==25.1.0",
    "pyjwt==2.10.1",
]

[project.optional-dependencies]
lint = ['ruff==0.14.10', 'mypy==1.19.1']

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.mypy]
strict = true
warn_unreachable = true
show_column_numbers = true
show_error_context = true
check_untyped_defs = true
ignore_missing_imports = true
warn_no_return = true

files = ["src/",]
exclude = ["src/backend/infrastructure/database/alembic/"]

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "src/**/*.py"]
exclude = ["src/**/infrastructure/database/alembic/**/*.py"]

[tool.ruff.lint]
select = ['ALL']
ignore = [
    # Strange and obscure
    'D100',
    'D104',
    'D101',
    'D105',
    'D102',
    'RET504',
    'D103',
    'PLR0913',
    'S101',
    'EM101',
    'TRY003',
    'D107',
    'ARG002',
    'RUF001',
    'PLR0912',
    'C901',
    'PGH003',
    'RUF003',
    'TC002',
    'D205',
    'RUF002',
    'TC001',
    'FBT001',
    'FBT002',
    'SIM103',
    'LOG015',
    'TC003',
    # Does not work correctly
    'S104',
    'ANN401',
    'PLR2004',
    'D203',
    'D213',
    'F821',
    # Not applicable now
    'ARG001',
]

[project.scripts]
backend = "backend.bootstrap.cli:main"
